Refactor GUI Into Separate Package
