+++ view-simulation-results/src/main/java/org/vcell/N5/reduction/DataReductionManager.java
@@ -9,6 +9,7 @@ import org.vcell.N5.UI.ControlButtonsPanel;
 import org.vcell.N5.UI.MainPanel;
 import org.vcell.N5.reduction.DTO.RangeOfImage;
 import org.vcell.N5.reduction.DTO.ReducedData;
+import org.vcell.N5.reduction.GUI.DataReductionGUI;
 import org.vcell.N5.retrieving.SimLoadingListener;
 import org.vcell.N5.retrieving.SimResultsLoader;
