+++ view-simulation-results/src/main/java/org/vcell/N5/reduction/GUI/DataReductionGUI.java
@@ -1,4 +1,4 @@
-package org.vcell.N5.reduction;
+package org.vcell.N5.reduction.GUI;
 
 import ij.ImagePlus;
 import ij.WindowManager;
