+++ view-simulation-results/src/main/java/org/vcell/N5/reduction/GUI/NormalizeGUI.java
@@ -1,4 +1,4 @@
-package org.vcell.N5.reduction;
+package org.vcell.N5.reduction.GUI;
 
 import org.vcell.N5.UI.HintTextField;
