+++ view-simulation-results/src/main/java/org/vcell/N5/reduction/ReductionCalculations.java
@@ -5,6 +5,7 @@ import ij.ImagePlus;
 import ij.gui.Roi;
 import org.vcell.N5.reduction.DTO.RangeOfImage;
 import org.vcell.N5.reduction.DTO.ReducedData;
+import org.vcell.N5.reduction.GUI.SelectMeasurements;
 
 import java.util.ArrayList;
 import java.util.HashMap;
